Refactored Path Sum III

+++ binary tree/PathSumIII.java
@@ -9,7 +9,7 @@
  * - The path must go downward (parent → child nodes).
  *
  * Approach:
- * 1. Use prefix sum + HashMap to track the cumulative sums.
+ * 1. Use prefix sum + HashMap to track cumulative sums as long.
  * 2. For each node, compute current cumulative sum.
  * 3. If (currSum - targetSum) exists in map, it indicates a valid path ending at current node.
  * 4. Recurse left and right, backtracking after recursion.
@@ -17,6 +17,8 @@
  * Time Complexity: O(n)
  * Space Complexity: O(n)
  */
+
+import java.util.HashMap;
 
 class TreeNode {
     int val;
@@ -36,33 +38,27 @@ class TreeNode {
     }
 }
 
-import java.util.HashMap;
-
 public class PathSumIII {
 
     public int pathSum(TreeNode root, int targetSum) {
-        HashMap<Integer, Integer> prefix = new HashMap<>();
-        prefix.put(0, 1); // base case
-        return dfs(root, 0, targetSum, prefix);
+        HashMap<Long, Integer> prefix = new HashMap<>();
+        prefix.put(0L, 1); // base case
+        return dfs(root, 0L, targetSum, prefix);
     }
 
-    private int dfs(TreeNode node, int currSum, int target, HashMap<Integer, Integer> prefix) {
+    private int dfs(TreeNode node, long currSum, int target, HashMap<Long, Integer> prefix) {
         if (node == null) return 0;
 
         currSum += node.val;
 
-        // Number of valid paths ending at current node
         int res = prefix.getOrDefault(currSum - target, 0);
 
-        // Update prefix map
         prefix.put(currSum, prefix.getOrDefault(currSum, 0) + 1);
 
-        // Recurse left and right
         res += dfs(node.left, currSum, target, prefix);
         res += dfs(node.right, currSum, target, prefix);
 
-        // Backtrack
-        prefix.put(currSum, prefix.get(currSum) - 1);
+        prefix.put(currSum, prefix.get(currSum) - 1); // backtrack
 
         return res;
     }
